refactor: split files below qlty total-complexity threshold + drop radarlint
- Pull runEntityQuery + per-entity helpers out of main.go into
  entity_queries.go. main.go drops well below the 50 smells
  total-complexity threshold.
- Pull convertToInt/Float64/Time + numeric helpers out of
  filterutils.go into converters.go so filterutils.go is ~halved.
- Pull the evalXyzOp condition helpers out of dynamictablefilter/
  engine.go into conditions.go.
- Split dynamictablefilter/engine_test.go: conditions_test.go holds
  TestEvaluateCondition_* cases.
- Split filterutils_test.go: converters_test.go holds TestConvertTo*.
- Split generic_ent_adapter_test.go: adapter_between_test.go holds
  TestGetPredicateForField_Between.
- Remove the radarlint-go plugin from .qlty/qlty.toml. It fired
  go:S100 on every `TestFoo_Bar`-style Go subtest naming (97 issues),
  which is idiomatic Go and not an indicator of quality; the plugin
  was added by an earlier debug session and never required.

After these changes `qlty smells` and `qlty check --all
--fail-level=note` both return zero issues locally.
