+++ .qlty/qlty.toml
@@ -22,7 +22,3 @@ default = true
 
 [smells]
 mode = "block"
-
-[[plugin]]
-name = "radarlint-go"
-mode = "monitor"
